ferdig med gui
